+++ src/main/scala/org/podval/xml/XmlError.scala
@@ -1,3 +1,4 @@
 package org.podval.xml
 
-final class XmlError(message: String) extends Exception(message)
+final class XmlError(message: String, cause: Throwable) extends Exception(message, cause):
+  def this(message: String) = this(message, null)